new loplugin: convertlong
merge the droplong and convertuintptr into one new plugin.
Limit the analysis to looking at var decl's, since that seems to be
safest proposition, even if that too needs some careful analysis.

+++ bridges/source/cpp_uno/gcc3_linux_x86-64/except.cxx
@@ -56,7 +56,7 @@ static OUString toUNOname( char const * p )
     while (*p != 'E')
     {
         // read chars count
-        long n = (*p++ - '0');
+        int n = (*p++ - '0');
         while ('0' <= *p && '9' >= *p)
         {
             n *= 10;
